separate CurrentAttempt from TaskDetailsToolbar
