+++ frontend/src/components/tasks/TaskDetails/CollapsibleToolbar.tsx
@@ -3,11 +3,7 @@ import { Button } from '@/components/ui/button.tsx';
 import { ChevronDown, ChevronUp } from 'lucide-react';
 import { TaskDetailsToolbar } from '@/components/tasks/TaskDetailsToolbar.tsx';
 
-type Props = {
-  projectHasDevScript?: boolean;
-};
-
-function CollapsibleToolbar({ projectHasDevScript }: Props) {
+function CollapsibleToolbar() {
   const [isHeaderCollapsed, setIsHeaderCollapsed] = useState(false);
 
   return (
@@ -29,9 +25,7 @@ function CollapsibleToolbar({ projectHasDevScript }: Props) {
           )}
         </Button>
       </div>
-      {!isHeaderCollapsed && (
-        <TaskDetailsToolbar projectHasDevScript={projectHasDevScript} />
-      )}
+      {!isHeaderCollapsed && <TaskDetailsToolbar />}
     </div>
   );
 }
